+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.3.0
+
+* Start using [`onhaschange`](https://developer.mozilla.org/en/docs/Web/API/WindowEventHandlers/onhashchange) API as a default for hash based routing.
+* Using [karma](https://www.npmjs.com/package/karma) as a test runner.
+
 ## 3.2.1
 
 Fixing the case where route handlers are not fired when only the query string part of the URL is changed (#53)